perf: bf16 Newton-Schulz iterations in Muon optimizer
Keep NS iterations in the parameter's dtype (bf16) instead of upcasting
to fp32. This uses bf16 tensor cores (989 TFLOPS) vs TF32 (495 TFLOPS),
halving optimizer step time for large models.

Only the norm computation upcasts to fp32 to avoid overflow.

Benchmarks (single H100, Muon optimizer, 100 steps):
  4B: 29 → 34 c/s (+17%), MFU 41.6% → 48.1%
  7B: 12 → 16 c/s (+33%), MFU 28.9% → 38.6%

+++ src/state/tx/optim.py
@@ -19,12 +19,14 @@ def _orthogonalize_update(matrix: torch.Tensor, steps: int, eps: float) -> torch
         raise ValueError(f"Expected a 2D update matrix, got shape {tuple(matrix.shape)}")
 
     transposed = False
-    x = matrix.float()
+    # Keep in input dtype (bf16) for 2× faster tensor-core matmuls vs fp32/TF32.
+    # Only the norm computation upcasts to fp32 to avoid overflow.
+    x = matrix
     if x.shape[0] > x.shape[1]:
         x = x.transpose(0, 1)
         transposed = True
 
-    x = x / (torch.linalg.norm(x) + eps)
+    x = x / (torch.linalg.norm(x.float()) + eps)
     a, b, c = 3.4445, -4.7750, 2.0315
     for _ in range(max(1, steps)):
         gram = x @ x.transpose(0, 1)
